pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.6",
    "google-genai>=0.2.2",
    "httpx>=0.27.2",
    "ipython>=8.30.0",
    "logfire[asyncpg]>=2.7.1",
    "mypy>=1.13.0",
    "polars>=1.17.1",
    "pydantic-ai>=0.0.13",
    "python-dotenv>=1.0.1",
    "supabase>=2.10.0",
    "tavily-python>=0.5.0",
    "uvicorn>=0.34.0",
]
